requirements.txt
# requirements.txt

streamlit
pandas
numpy
plotly
scikit-learn
tensorflow
python-dotenv
sqlalchemy
psycopg2-binary
PyMySQL
prophet